auth_service/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
parso==0.8.3
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
